requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-MySQLdb==1.0.1
reportlab==4.1.0
#ngrok http 5000